add registers
enums don't translate so well from c. This is pretty verbose to use
and could, alternatively, be achieved using a mapping function from
the enum to the right value in the right format. Since we've only got
the one piece of information that we want to associate with the variant
we'll use an integer tag value.